// src/types/Tags.ts
enum Tags {
  ARCADE_CABINET = 'Arcade-cabinet',
  CHALLENGING = 'Challenging',
  FIRST_PERSON = 'First-person',
  NO_PVE = 'No PvE',
  NOSTALGIA = 'Nostalgia',
  OVERWHELMINGLY_POSITIVE = 'Overwhelmingly positive (Steam)',
  PVP = 'Has PvP',
  SCI_FI = 'Sci-fi',
  // SHOOTER_SLASHER = 'Shooter/Slasher',
  THIRD_PERSON = 'Third-person',
  TOP_DOWN = 'Top-down',
}

export default Tags;